#57 fix dump function
