+++ javascript-console-repo/src/main/resources/de/fme/jsconsole/jsconsole-pre-roll-script.js
@@ -4,5 +4,5 @@ function print(obj) {
 };
 
 function dump(obj){
-    jsconsole.addDump(obj);
+    dumpService.addDump(obj);
 };
